requirements.txt
langchain
google-generativeai
streamlit
langchain-google-genai
faiss-cpu
fastapi
uvicorn
pydantic
python-dotenv
python-multipart